[#39] - Window Extension으로 파일 분리
